Fix build with system libheif >= 1.20.0
A change occurred in the libheif API for versions 1.20.0 and 1.20.1,
which is incompatible with versions 1.19 and >= 1.20.2

Issue https://github.com/strukturag/libheif/issues/1566

This patch fixes the compilation of the addon that uses the system versions of libheif.

The fix has been tested on Gentoo Linux with media-libs/libheif 1.19.8, 1.20.1, and 1.20.2.